requirements.txt
ib_insync>=0.9.80
pandas>=1.3.5
numpy>=1.21.6
jinja2>=3.0.3
matplotlib>=3.5.3
seaborn>=0.12.0
python-dotenv>=0.19.0
loguru>=0.5.3
backtrader>=1.9.76

# Web application dependencies
flask>=2.3.0
flask-cors>=3.0.10
gunicorn>=20.1.0
werkzeug>=2.3.0 